use decimal timestamp instead of int

+++ orderbook/order.py
@@ -1,5 +1,6 @@
-from decimal import * 
+from decimal import *
 import time, random
+
 
 class Order(object):
     '''
@@ -9,7 +10,7 @@ class Order(object):
     existing Order.
     '''
     def __init__(self, quote, order_list):
-        self.timestamp = int(quote['timestamp']) # integer representing the timestamp of order creation
+        self.timestamp = Decimal(quote['timestamp']) # decimal representing the timestamp of order creation
         self.quantity = Decimal(quote['quantity']) # decimal representing amount of thing - can be partial amounts
         self.price = Decimal(quote['price']) # decimal representing price (currency)
         self.order_id = int(quote['order_id'])
